Wall Detection
Now has a proper way of handling wall detection. Using the WallCheck() method, a raycast is deployed every time the player attempts to move and detects whether or not a wall is in front of the player. If an object with the tag "Wall" is in the desired direction of movement, the player will not move. (Also made some minor tweaks/adjustments to neatness + orderliness of code)